quickreserve-project/reminder_scheduler/requirements.txt
functions-framework
requests
google-cloud-firestore
pytz